+++ go.mod
@@ -3,7 +3,7 @@ module github.com/Namecheap-Ecosystem/terraform-provider-namecheap
 go 1.15
 
 require (
-	github.com/Namecheap-Ecosystem/go-namecheap v0.0.0-20210213023153-2a164e62be82
+	github.com/Namecheap-Ecosystem/go-namecheap v0.0.0-20210310202035-1b86067b819e
 	github.com/hashicorp/hcl/v2 v2.8.2 // indirect
 	github.com/hashicorp/terraform-plugin-sdk/v2 v2.3.0
 	github.com/stretchr/testify v1.5.1 // indirect
