+++ namecheap/resource_domain_dns.go
@@ -9,7 +9,6 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 )
 
 // API doc: https://www.namecheap.com/support/api/methods/
@@ -21,26 +20,11 @@ func resourceDomainDNS() *schema.Resource {
 				Type:        schema.TypeString,
 				Required:    true,
 			},
-			"hosts": {
-				Type:     schema.TypeSet,
-				Required: true,
-				Elem: &schema.Resource{
-					Schema: map[string]*schema.Schema{
-						"type": {
-							Type:         schema.TypeString,
-							Required:     true,
-							ValidateFunc: validation.StringInSlice([]string{"A", "AAAA", "ALIAS", "CAA", "CNAME", "MX", "MXE", "NS", "TXT", "URL", "URL301", "FRAME"}, false),
-						},
-						"address": {
-							Type:     schema.TypeString,
-							Required: true,
-						},
-						"ttl": {
-							Type:     schema.TypeInt,
-							Required: true,
-						},
-					},
-				},
+			"nameservers": {
+				Description: "List of custom nameservers to be associated with the domain name",
+				Type:        schema.TypeList,
+				Required:    true,
+				Elem:        &schema.Schema{Type: schema.TypeString},
 			},
 		},
 		CreateContext: resourceDomainDNSCreate,
@@ -53,17 +37,31 @@ func resourceDomainDNS() *schema.Resource {
 	}
 }
 
+func resourceDomainDNSCreate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+	c := meta.(*namecheap.Client)
+
+	sld, tld := parseDomain(d.Get("domain").(string))
+	nameservers := strings.Join(expandStringListFromSetSchema(d.Get("nameservers").(*schema.Set)), ",")
+	res, err := c.DomainDNSSetCustom(sld, tld, nameservers)
+	if err != nil {
+		return diag.FromErr(err)
+	}
+
+	d.SetId(res.Domain)
+	return resourceDomainDNSRead(ctx, d, meta)
+}
+
 func resourceDomainDNSRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	c := meta.(*namecheap.Client)
 	var diags diag.Diagnostics
 
 	sld, tld := parseDomain(d.Id())
-	res, err := c.DomainsDNSGetHosts(sld, tld)
+	res, err := c.DomainDNSGetList(sld, tld)
 	if err != nil {
 		return diag.FromErr(err)
 	}
 
-	if err := d.Set("hosts", flattenHosts(res.Hosts)); err != nil {
+	if err := d.Set("nameservers", res.Nameservers); err != nil {
 		return diag.FromErr(err)
 	}
 
@@ -71,34 +69,12 @@ func resourceDomainDNSRead(ctx context.Context, d *schema.ResourceData, meta int
 	return diags
 }
 
-func resourceDomainDNSCreate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
-	c := meta.(*namecheap.Client)
-
-	sld, tld := parseDomain(d.Get("domain").(string))
-	hosts := expandHosts(d)
-	result, err := c.DomainDNSSetHosts(sld, tld, hosts)
-	if err != nil {
-		return diag.FromErr(err)
-	}
-
-	d.SetId(result.Domain)
-	return resourceDomainDNSRead(ctx, d, meta)
-}
-
 func resourceDomainDNSUpdate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	c := meta.(*namecheap.Client)
 
 	sld, tld := parseDomain(d.Id())
-	res, err := c.DomainsDNSGetHosts(sld, tld)
-	if err != nil {
-		return diag.FromErr(err)
-	}
-
-	if d.HasChange("roles") {
-		res.Hosts = expandHosts(d)
-	}
-
-	_, err = c.DomainDNSSetHosts(sld, tld, res.Hosts)
+	nameservers := strings.Join(expandStringListFromSetSchema(d.Get("nameservers").(*schema.Set)), ",")
+	_, err := c.DomainDNSSetCustom(sld, tld, nameservers)
 	if err != nil {
 		return diag.FromErr(err)
 	}
@@ -111,8 +87,7 @@ func resourceDomainDNSDelete(ctx context.Context, d *schema.ResourceData, meta i
 	var diags diag.Diagnostics
 
 	sld, tld := parseDomain(d.Id())
-	_, err := c.DomainsDNSGetHosts(sld, tld)
-	if err != nil {
+	if _, err := c.DomainDNSSetDefault(sld, tld, []namecheap.DomainDNSHost{}); err != nil {
 		return diag.FromErr(err)
 	}
 
@@ -122,47 +97,8 @@ func resourceDomainDNSDelete(ctx context.Context, d *schema.ResourceData, meta i
 
 func resourceDomainDNSImportState(ctx context.Context, d *schema.ResourceData, meta interface{}) ([]*schema.ResourceData, error) {
 	if diags := resourceDomainDNSRead(ctx, d, meta); diags.HasError() {
-		return nil, fmt.Errorf("failed to import domain nameserver")
+		return nil, fmt.Errorf("failed to import domain dns")
 	}
 
 	return []*schema.ResourceData{d}, nil
-}
-
-func parseDomain(domain string) (string, string) {
-	elements := strings.Split(domain, ".")
-	return elements[0], elements[1]
-}
-
-func expandHosts(d *schema.ResourceData) []namecheap.DomainDNSHost {
-	var hosts []namecheap.DomainDNSHost
-
-	if v, ok := d.GetOk("hosts"); ok {
-		if hs := v.(*schema.Set); hs.Len() > 0 {
-			hosts = make([]namecheap.DomainDNSHost, hs.Len())
-
-			for k, r := range hs.List() {
-				hostMap := r.(map[string]interface{})
-				hosts[k] = namecheap.DomainDNSHost{
-					Type:    hostMap["type"].(string),
-					TTL:     hostMap["ttl"].(int),
-					Address: hostMap["address"].(string),
-				}
-			}
-		}
-	}
-
-	return hosts
-}
-
-func flattenHosts(hosts []namecheap.DomainDNSHost) interface{} {
-	hostList := make([]interface{}, len(hosts))
-	for i, v := range hosts {
-		hostList[i] = map[string]interface{}{
-			"type":    v.TTL,
-			"ttl":     v.TTL,
-			"address": v.Address,
-		}
-	}
-
-	return hostList
 }
